move parsing of gocode suggestions and generating autocomplete-plus suggestions into separate file
